src/requirements.txt
asgiref==3.11.0
attrs==25.4.0
bandit==1.9.3
black==26.1.0
cfgv==3.5.0
click==8.3.1
distlib==0.4.0
Django==6.0.1
django-environ==0.12.0
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-stubs==3.16.8
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
filelock==3.20.3
flake8==7.3.0
gunicorn==25.0.1
identify==2.6.16
inflection==0.5.1
iniconfig==2.3.0
isort==7.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
librt==0.7.8
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy==1.19.1
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==26.0
pathspec==1.0.4
pep8-naming==0.15.1
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
psycopg==3.3.2
psycopg-binary==3.3.2
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-django==4.11.1
python-dotenv==1.2.1
pytokens==0.4.1
PyYAML==6.0.3
referencing==0.37.0
rich==14.3.2
rpds-py==0.30.0
sqlparse==0.5.5
stevedore==5.6.0
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
uritemplate==4.2.0
virtualenv==20.36.1
